requirements.txt
# Python dependencies for Legacy Donation backend
# Install with: pip install -r requirements.txt

# MySQL database connector
mysql-connector-python>=8.1.0

# Flask web framework for API endpoints
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0

# Database utilities and ORM
SQLAlchemy>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation and serialization
marshmallow>=3.20.0

# Logging and monitoring
APScheduler>=3.10.0

# JSON handling and validation
jsonschema>=4.19.0

# CORS handling
Flask-Cors>=4.0.0

# Development dependencies
pytest>=7.4.0
pytest-flask>=1.2.0
python-dotenv>=1.0.0